fix: Move functions out of Form
